twilio_test/requirements.txt
twilio==8.12.0
python-dotenv==1.0.1 